refactor(jsonkeys): extract per-key JSON merge from claude adapter
Identical behavior, shared package. Claude re-exports MergeKeys for
internal callers; opencode and future adapters import jsonkeys directly.
